Change the values for the magic
The previous version of the spec used the values for the
memcached binary protocol for the magic type of the packets, but
it changed the format of the header so it is no longer compatible
with the memcached binary protocol (the datatype byte is
scheduled to be utilized, so we can't use that for something
else). This makes it natural to change the magic to its own
series.

In addition to the problem with the incompatible header layout
the proposed opcodes was in conflict with the existing opcodes
used by the memcached binary protocol.